Add files via upload
This was a sentiment project I worked on with two other students. We analyzed Beatles lyrics to see if we could predict who the author of a song was. I focused on analyzing John Lennon's and Paul McCartney's lyrics and found some interests sentiment results. Along with the sentiment analysis I had some fun creating some word clouds.